Refactor and Ordering
